requirements.txt
aiogram~=3.19.0
python-dotenv~=1.1.0
py3xui~=0.3.6
aiosqlite~=0.21.0
SQLAlchemy~=2.0.40
shortuuid~=1.0.13